Create IO abstraction library for fs and network syscalls
This will allow us to:

1. Write code in an easy synchronous style, despite the underlying IO syscalls
being non-blocking and async through io_uring... a miracle of Zig's colorless
async/await (https://www.youtube.com/watch?v=zeLToGnjIUM). For example, calling
sleep() will not block the process, only the call site, the rest of the event
loop will keep running, and this all works for a single thread, without any
multithreading, to support our thread-per-core design.

2. Write fault-injection shims to simulate all kinds of network and file
system errors when running integration tests.

3. Add cross-platform IO support for Windows (IOCP), FreeBSD/Mac (kqueue)
and older versions of Linux (epoll) if necessary down the line.

4. Rely on the compiler to ensure that all syscall errors are handled
correctly, without missing any, since this also converts the `errno` returned
from io_uring to a Zig error type, using the same Zig errors as the std lib.

This adds support for the following syscalls through io_uring:

accept() a TCP connection as a server
connect() to a TCP server as a client

recv() into a buffer from a TCP connection
send() from a buffer to a TCP connection

openat() a file descriptor
close() a file descriptor

read() into a buffer from a file descriptor
write() from a buffer to a file descriptor

fsync() buffered data for the entire file system of a fd to the disk platter

sleep() for a number of milliseconds

As a bonus, this handles and retries signal interrupt errors when calling into
io_uring, as well as submission and completion queue overcommit errors.